requirements.txt
transformers==4.36.2
sentence-transformers==2.2.2
llama-index==0.9.30
openai==1.7.1
langchain==0.1.0
newspaper3k==0.2.8
lxml==5.1.0
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.2
rank-bm25==0.2.2
nltk==3.8.1
gradio==4.14.0